DRTII-60 Refactor common code to interface
